+++ zonerole_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/keycardai/keycard-go/option"
 )
 
-func TestZoneResourceNewWithOptionalParams(t *testing.T) {
+func TestZoneRoleNewWithOptionalParams(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -28,23 +28,14 @@ func TestZoneResourceNewWithOptionalParams(t *testing.T) {
 		option.WithClientID("My Client ID"),
 		option.WithClientSecret("My Client Secret"),
 	)
-	_, err := client.Zones.Resources.New(
+	_, err := client.Zones.Roles.New(
 		context.TODO(),
 		"zoneId",
-		keycard.ZoneResourceNewParams{
-			Identifier:                "x",
-			Name:                      "x",
-			ApplicationID:             keycard.String("application_id"),
-			ApplicationType:           keycard.ZoneResourceNewParamsApplicationTypeNative,
-			CredentialLifetimeSeconds: keycard.Int(60),
-			CredentialProviderID:      keycard.String("credential_provider_id"),
-			Description:               keycard.String("description"),
-			Metadata: keycard.MetadataParam{
-				DocsURL: keycard.String("https://example.com"),
-				IconURL: keycard.String("https://example.com"),
+		keycard.ZoneRoleNewParams{
+			RoleCreate: keycard.RoleCreateParam{
+				Identifier:  "identifier",
+				Description: keycard.String("description"),
 			},
-			Prefix: keycard.Bool(true),
-			Scopes: []string{"x"},
 		},
 	)
 	if err != nil {
@@ -56,7 +47,7 @@ func TestZoneResourceNewWithOptionalParams(t *testing.T) {
 	}
 }
 
-func TestZoneResourceGet(t *testing.T) {
+func TestZoneRoleGet(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -71,10 +62,10 @@ func TestZoneResourceGet(t *testing.T) {
 		option.WithClientID("My Client ID"),
 		option.WithClientSecret("My Client Secret"),
 	)
-	_, err := client.Zones.Resources.Get(
+	_, err := client.Zones.Roles.Get(
 		context.TODO(),
-		"id",
-		keycard.ZoneResourceGetParams{
+		"roleId",
+		keycard.ZoneRoleGetParams{
 			ZoneID: "zoneId",
 		},
 	)
@@ -87,7 +78,7 @@ func TestZoneResourceGet(t *testing.T) {
 	}
 }
 
-func TestZoneResourceUpdateWithOptionalParams(t *testing.T) {
+func TestZoneRoleUpdateWithOptionalParams(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -102,24 +93,14 @@ func TestZoneResourceUpdateWithOptionalParams(t *testing.T) {
 		option.WithClientID("My Client ID"),
 		option.WithClientSecret("My Client Secret"),
 	)
-	_, err := client.Zones.Resources.Update(
+	_, err := client.Zones.Roles.Update(
 		context.TODO(),
-		"id",
-		keycard.ZoneResourceUpdateParams{
-			ZoneID:                    "zoneId",
-			ApplicationID:             keycard.String("application_id"),
-			ApplicationType:           keycard.ZoneResourceUpdateParamsApplicationTypeNative,
-			CredentialLifetimeSeconds: keycard.Int(60),
-			CredentialProviderID:      keycard.String("credential_provider_id"),
-			Description:               keycard.String("description"),
-			Identifier:                keycard.String("x"),
-			Metadata: keycard.MetadataUpdateParam{
-				DocsURL: keycard.String("https://example.com"),
-				IconURL: keycard.String("https://example.com"),
+		"roleId",
+		keycard.ZoneRoleUpdateParams{
+			ZoneID: "zoneId",
+			RoleUpdate: keycard.RoleUpdateParam{
+				Description: keycard.String("description"),
 			},
-			Name:   keycard.String("x"),
-			Prefix: keycard.Bool(true),
-			Scopes: []string{"x"},
 		},
 	)
 	if err != nil {
@@ -131,7 +112,7 @@ func TestZoneResourceUpdateWithOptionalParams(t *testing.T) {
 	}
 }
 
-func TestZoneResourceListWithOptionalParams(t *testing.T) {
+func TestZoneRoleListWithOptionalParams(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -146,26 +127,17 @@ func TestZoneResourceListWithOptionalParams(t *testing.T) {
 		option.WithClientID("My Client ID"),
 		option.WithClientSecret("My Client Secret"),
 	)
-	_, err := client.Zones.Resources.List(
+	_, err := client.Zones.Roles.List(
 		context.TODO(),
 		"zoneId",
-		keycard.ZoneResourceListParams{
-			After:                keycard.String("x"),
-			Before:               keycard.String("x"),
-			CredentialProviderID: keycard.String("credentialProviderId"),
-			Expand: keycard.ZoneResourceListParamsExpandUnion{
-				OfZoneResourceListsExpandString: keycard.String("total_count"),
-			},
-			FilterIdentifier: keycard.ZoneResourceListParamsFilterIdentifierUnion{
-				OfString: keycard.String("string"),
-			},
-			FilterOwnerType: keycard.ZoneResourceListParamsFilterOwnerTypePlatform,
-			FilterTraits: keycard.ZoneResourceListParamsFilterTraitsUnion{
-				OfString: keycard.String("string"),
+		keycard.ZoneRoleListParams{
+			After:  keycard.String("x"),
+			Before: keycard.String("x"),
+			Expand: keycard.ZoneRoleListParamsExpandUnion{
+				OfZoneRoleListsExpandString: keycard.String("total_count"),
 			},
 			Identifier: keycard.String("identifier"),
 			Limit:      keycard.Int(1),
-			Slug:       keycard.String("slug"),
 		},
 	)
 	if err != nil {
@@ -177,7 +149,7 @@ func TestZoneResourceListWithOptionalParams(t *testing.T) {
 	}
 }
 
-func TestZoneResourceDelete(t *testing.T) {
+func TestZoneRoleDelete(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -192,10 +164,10 @@ func TestZoneResourceDelete(t *testing.T) {
 		option.WithClientID("My Client ID"),
 		option.WithClientSecret("My Client Secret"),
 	)
-	err := client.Zones.Resources.Delete(
+	err := client.Zones.Roles.Delete(
 		context.TODO(),
-		"id",
-		keycard.ZoneResourceDeleteParams{
+		"roleId",
+		keycard.ZoneRoleDeleteParams{
 			ZoneID: "zoneId",
 		},
 	)