+++ zoneuser_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/keycardai/keycard-go/option"
 )
 
-func TestZoneUserGet(t *testing.T) {
+func TestZoneUserGetWithOptionalParams(t *testing.T) {
 	t.Skip("Mock server tests are disabled")
 	baseURL := "http://localhost:4010"
 	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
@@ -33,6 +33,43 @@ func TestZoneUserGet(t *testing.T) {
 		"id",
 		keycard.ZoneUserGetParams{
 			ZoneID: "zoneId",
+			Expand: keycard.ZoneUserGetParamsExpandUnion{
+				OfZoneUserGetsExpandString: keycard.String("role-assignments"),
+			},
+			RoleSource: keycard.ZoneUserGetParamsRoleSourceUser,
+		},
+	)
+	if err != nil {
+		var apierr *keycard.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestZoneUserUpdateWithOptionalParams(t *testing.T) {
+	t.Skip("Mock server tests are disabled")
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := keycard.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("My API Key"),
+		option.WithClientID("My Client ID"),
+		option.WithClientSecret("My Client Secret"),
+	)
+	_, err := client.Zones.Users.Update(
+		context.TODO(),
+		"id",
+		keycard.ZoneUserUpdateParams{
+			ZoneID:     "zoneId",
+			Identifier: keycard.String("x"),
+			Status:     keycard.ZoneUserUpdateParamsStatusActive,
 		},
 	)
 	if err != nil {
@@ -71,6 +108,9 @@ func TestZoneUserListWithOptionalParams(t *testing.T) {
 			FilterEmail: keycard.ZoneUserListParamsFilterEmailUnion{
 				OfString: keycard.String("dev@stainless.com"),
 			},
+			FilterGroups: keycard.ZoneUserListParamsFilterGroupsUnion{
+				OfString: keycard.String("string"),
+			},
 			FilterID: keycard.ZoneUserListParamsFilterIDUnion{
 				OfString: keycard.String("string"),
 			},
@@ -87,7 +127,39 @@ func TestZoneUserListWithOptionalParams(t *testing.T) {
 			QuerySubject: keycard.ZoneUserListParamsQuerySubjectUnion{
 				OfString: keycard.String("x"),
 			},
-			Sort: keycard.String("-authenticated_at,\t\r\r \tauthenticated_at,\n\f\t\f\ncreated_at"),
+			RoleSource: keycard.ZoneUserListParamsRoleSourceUser,
+			Sort:       keycard.String("-authenticated_at,\t\r\r \tauthenticated_at,\n\f\t\f\ncreated_at"),
+		},
+	)
+	if err != nil {
+		var apierr *keycard.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestZoneUserDelete(t *testing.T) {
+	t.Skip("Mock server tests are disabled")
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := keycard.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("My API Key"),
+		option.WithClientID("My Client ID"),
+		option.WithClientSecret("My Client Secret"),
+	)
+	err := client.Zones.Users.Delete(
+		context.TODO(),
+		"id",
+		keycard.ZoneUserDeleteParams{
+			ZoneID: "zoneId",
 		},
 	)
 	if err != nil {